+++ changelog.md
@@ -1,3 +1,9 @@
+## 20.04.2023, Thu
+
+- 20:30 - Frontend: added loading indicator on "Refresh" button (issue #35).
+- 20:45 - Frontend: added loading indicator on "Save changes" button (issue #35).
+- 21:00 - Frontend: added loading indicator on the Bookmarks page on initial data fetch (issue #35).
+
 ## 19.04.2023, Wed
 
 - 22:30 - Frontend: added loading indicator when adding new bookmark (issue #35).